sql: move CharSetHandler to sql package
